requirements.txt
elasticsearch==9.1.0
openai==1.101.0
python_docx==1.1.2
tqdm==4.66.5
wxauto==39.1.15
